Add scripts/repackage-common.sh to improve update process
This script makes it easier to use currysrc to repackage source files
by removing the need for each repository to create its own custom
script for invoking currysrc. Instead they simply have to specify a
few fields, include this script to do the repackaging, and then, if
necessary do any post repackaging cleanup.

This script also checks that every change that it requested to be
made is actually made and fails otherwise. This ensures that an update
does not accidentally remove a method that is part of the hiddenapi and
which could cause app compatibility issues.

Bug: 119243026
Test: external/libphonenumber/srcgen/generate_android_src.sh
Change-Id: Ib60d6e035e65d99ff5b90e35e9a4aff43023311b